requirements.txt
allure-pytest==2.8.13
pytest==5.4.1
pytest-html
selenium==3.141.0